fix: selection of next responsible user id
There was a bug in `getNextResponsibleUserId`. For `lastAssignments`, we should have fetched the last $AMOUNT_OF_USERS
assignments, distinct by user, but we just fetched the last $AMOUNT_OF_USERS, so that they were all from the same user which broke
the selection logic. We now rely on the `assignmentOrdinal` instead, which is now not nullable anymore.